Add some comments, remove some memory leaks.
Trying to make the code a bit more readable.

I bellieve I have removed the worst memory leaks, and also made the code
a little more sane.

The video-data parsing functions are not working correctly at the
moment.
Because of this, I don't think the binary data that is streamed out of the
/dev/videoX file is correctly formated as video frames, so don't expect it to work with
mplayer et.al.

I have not tested the V4L2 API, but I guess it's still not
working!

The only thing that works so far, is the ability to cat data from
/dev/videoX to a file, and see that the ISOC data is transmitted.

-- WARNING, THIS DRIVER IS EXTREMLY UNSTABLE, AND SHOULD NOT BE RUN BY ANYONE...