Improve frame sync
The parse_data function is now trying to read one line at a time,
instead of copying the frame byte-by-byte from the scratch buffer.

Known bugs:
The capture will on random occations start at the wrong field.
This will cause the driver to skip every other frame, and you will end
up with 12 FPS on a PAL capture. (And probably 15 FPS on NTSC, but I
have not checked this)

Workaround:
cancel runing capture, and try to start a new one.